Stop clearing the cache in github_repo_fetcher_spec.rb
The way these tests were written meant that the cache got cleared
when the tests were run, so if we'd just gone through the
expensive (many requests, multiple minutes) build process, we'd
have to go through it allllll again to get requests in the cache.

We can now build the project from scratch (i.e. make lots of HTTP
requests), then run the test suite, then build the project again
(this time with no HTTP requests, as the requests are cached), and
it all works nicely.